attackers attack and defenders perish

+++ Assets/Scripts/Health.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 public class Health : MonoBehaviour
 {
     [SerializeField]
-    int health = 100;
+    float health = 100;
 
     [SerializeField]
     GameObject deathVFX;
 
-    public void DealDamage(int damage)
+    public void DealDamage(float damage)
     {
         health -= damage;
